(PC-15858)[API] refactor: share change email models
